move new DB functions into separate file
