+++ crates/database/src/orders.rs
@@ -689,78 +689,6 @@ pub async fn single_full_order_with_quote(
     sqlx::query_as(QUERY).bind(uid).fetch_optional(ex).await
 }
 
-/// A fast-path order's signed fields and interactions, joined to its quote's
-/// winning solution and recorded fill. Only what the driver needs to re-encode
-/// the settlement is selected; order metadata and quote data are left out.
-#[derive(Debug, sqlx::FromRow)]
-pub struct FastPathOrder {
-    pub uid: OrderUid,
-    pub owner: Address,
-    pub creation_timestamp: DateTime<Utc>,
-    pub sell_token: Address,
-    pub buy_token: Address,
-    pub sell_amount: BigDecimal,
-    pub buy_amount: BigDecimal,
-    pub valid_to: i64,
-    pub app_data: AppId,
-    pub kind: OrderKind,
-    pub partially_fillable: bool,
-    pub signature: Vec<u8>,
-    pub receiver: Option<Address>,
-    pub signing_scheme: SigningScheme,
-    pub sell_token_balance: SellTokenSource,
-    pub buy_token_balance: BuyTokenDestination,
-    pub pre_interactions: Vec<RawInteraction>,
-    pub post_interactions: Vec<RawInteraction>,
-    /// Contents of the order's `app_data` document (from the `app_data`
-    /// table). `None` when the full document was never uploaded.
-    pub full_app_data: Option<Vec<u8>>,
-    pub auction_id: AuctionId,
-    pub solution_id: BigDecimal,
-    pub solution_uid: i64,
-    pub solver: Address,
-    pub executed_sell: BigDecimal,
-    pub executed_buy: BigDecimal,
-    /// The quote auction's native prices (token, normalized price).
-    pub price_tokens: Vec<Address>,
-    pub price_values: Vec<BigDecimal>,
-}
-
-/// Recovers what's needed to settle `uid` out of competition in one query, or
-/// `None` when it is not a fast-path order (no persisted quote competition).
-#[instrument(skip_all)]
-pub async fn single_fast_path_order(
-    ex: &mut PgConnection,
-    uid: &OrderUid,
-) -> Result<Option<FastPathOrder>, sqlx::Error> {
-    #[rustfmt::skip]
-    const QUERY: &str = const_format::concatcp!(
-        "SELECT ",
-        "o.uid, o.owner, o.creation_timestamp, o.sell_token, o.buy_token, ",
-        "o.sell_amount, o.buy_amount, o.valid_to, o.app_data, o.kind, ",
-        "o.partially_fillable, o.signature, o.receiver, o.signing_scheme, ",
-        "o.sell_token_balance, o.buy_token_balance, ",
-        "array(SELECT (p.target, p.value, p.data) FROM interactions p",
-        " WHERE p.order_uid = o.uid AND p.execution = 'pre' ORDER BY p.index) AS pre_interactions, ",
-        "array(SELECT (p.target, p.value, p.data) FROM interactions p",
-        " WHERE p.order_uid = o.uid AND p.execution = 'post' ORDER BY p.index) AS post_interactions, ",
-        "ad.full_app_data AS full_app_data, ",
-        "oq.auction_id AS auction_id, ps.id AS solution_id, ps.uid AS solution_uid, ps.solver AS solver, ",
-        "pte.executed_sell AS executed_sell, pte.executed_buy AS executed_buy, ",
-        "ca.price_tokens AS price_tokens, ca.price_values AS price_values",
-        " FROM orders o",
-        " JOIN order_quotes oq ON oq.order_uid = o.uid",
-        " JOIN proposed_solutions ps ON ps.auction_id = oq.auction_id AND ps.is_winner",
-        " JOIN proposed_trade_executions pte",
-        " ON pte.auction_id = ps.auction_id AND pte.solution_uid = ps.uid AND pte.order_uid = o.uid",
-        " JOIN competition_auctions ca ON ca.id = oq.auction_id",
-        " LEFT JOIN app_data ad ON ad.contract_app_data = o.app_data",
-        " WHERE o.uid = $1",
-        " LIMIT 1",
-    );
-    sqlx::query_as(QUERY).bind(uid).fetch_optional(ex).await
-}
-
 #[instrument(skip_all)]
 pub async fn many_full_orders_with_quotes<'a>(
     ex: &'a mut PgConnection,
