+++ crates/database/src/solver_competition_v2.rs
@@ -355,94 +355,6 @@ async fn save_jit_orders(
     Ok(())
 }
 
-/// Because the final order uid is not known when we store the quote
-/// competition data we use `0x000...000` as a sentinel value.
-/// When an order gets placed referencing a quote competition this function
-/// replaces the placeholder value with the now final order uid.
-#[instrument(skip_all)]
-pub async fn finalize_quote_competition(
-    ex: &mut PgTransaction<'_>,
-    auction_id: AuctionId,
-    order_uid: OrderUid,
-) -> Result<(), sqlx::Error> {
-    const QUERY: &str = r#"
-WITH patch_te AS (
-    UPDATE proposed_trade_executions
-    SET order_uid = $1
-    WHERE auction_id = $2 AND order_uid = $3
-)
-UPDATE competition_auctions
-SET order_uids = ARRAY[$1]
-WHERE id = $2
-"#;
-    sqlx::query(QUERY)
-        .bind(order_uid)
-        .bind(auction_id)
-        .bind(OrderUid::default())
-        .execute(ex.deref_mut())
-        .await?;
-    Ok(())
-}
-
-#[derive(Debug, Clone, sqlx::FromRow)]
-pub struct FastPathBid {
-    pub solution_uid: i64,
-    pub executed_sell: BigDecimal,
-    pub executed_buy: BigDecimal,
-}
-
-/// Returns every proposed trade execution recorded against the fast-path
-/// order — one row per competing solver. Rows for JIT orders (which use
-/// different `order_uid`s) are naturally excluded.
-#[instrument(skip_all)]
-pub async fn fast_path_bids(
-    ex: &mut PgConnection,
-    auction_id: AuctionId,
-    order_uid: OrderUid,
-) -> Result<Vec<FastPathBid>, sqlx::Error> {
-    const QUERY: &str = r#"
-SELECT solution_uid, executed_sell, executed_buy
-FROM proposed_trade_executions
-WHERE auction_id = $1 AND order_uid = $2
-"#;
-    sqlx::query_as(QUERY)
-        .bind(auction_id)
-        .bind(order_uid)
-        .fetch_all(ex)
-        .await
-}
-
-/// Overwrites the executed amounts on every competing solver's bid for a
-/// fast-path order in a single query. Each row is matched by its own
-/// `solution_uid`, so different bids can be updated to different values.
-#[instrument(skip_all)]
-pub async fn apply_fees_to_fast_path_bids(
-    ex: &mut PgConnection,
-    auction_id: AuctionId,
-    order_uid: OrderUid,
-    bids: &[FastPathBid],
-) -> Result<(), sqlx::Error> {
-    if bids.is_empty() {
-        return Ok(());
-    }
-    let mut query_builder = QueryBuilder::new(
-        "UPDATE proposed_trade_executions AS pte SET executed_sell = v.executed_sell, \
-         executed_buy = v.executed_buy FROM (",
-    );
-    query_builder.push_values(bids.iter(), |mut b, bid| {
-        b.push_bind(bid.solution_uid)
-            .push_bind(&bid.executed_sell)
-            .push_bind(&bid.executed_buy);
-    });
-    query_builder.push(") AS v(solution_uid, executed_sell, executed_buy) WHERE pte.auction_id = ");
-    query_builder.push_bind(auction_id);
-    query_builder.push(" AND pte.order_uid = ");
-    query_builder.push_bind(order_uid);
-    query_builder.push(" AND pte.solution_uid = v.solution_uid");
-    query_builder.build().execute(ex).await?;
-    Ok(())
-}
-
 /// Deletes all competition rows associated with `auction_id` across
 /// `proposed_trade_executions`, `proposed_jit_orders`, `proposed_solutions`,
 /// and `competition_auctions`.
